+++ tests/Card/NextStatusTest.php
@@ -70,7 +70,7 @@ class NextStatusTest extends TestCase
         $card = new BanklyCard();
         $response = $card->nextStatus('12345678');
 
-        Http::assertSent(function ($request) {
+        Http::assertSent(function (\Illuminate\Http\Client\Request $request) {
             return Str::contains(
                 $request->url(),
                 '12345678/nextStatus'