+++ tests/Customer/CreateCustomerAccountTest.php
@@ -70,7 +70,7 @@ class CreateCustomerAccountTest extends TestCase
         $client = $this->getBanklyClient();
         $response = $client->createCustomerAccount('12345678909', $paymentAccount);
 
-        Http::assertSent(function ($request) {
+        Http::assertSent(function (\Illuminate\Http\Client\Request $request) {
             $body = collect($request->data());
 
             return $body['accountType'] === 'PAYMENT_ACCOUNT';